File preview Button Complete
The file preview button opens files in a new window. If that window is
already opened, the focus will be shifted to that open page and the page
will reload to show any changes. I also positioned the save and preview
buttons relative to the header and the editor div relative to the
buttons.