More work on validation,
So now it checks the files within a folder if there is one given in the
extension. For example, if they enter /Folder/first, it won't let them
create the new folder or file without first giving a valid ext to the
file. You also cannot make empty folders. Also altered the radio buttons
as file extensions to make things clearer to the user.